유효성검사 전

+++ src/main/java/com/example/netflixclone/repository/UserRepository.java
@@ -9,8 +9,8 @@ import com.example.netflixclone.model.User;
 
 // DAO
 // @Repository 생략 가능 => 자동으로 bean등록이 됨
-public interface UserRepository extends JpaRepository<User, Integer> {
-	// SELECT * FROM user WHERE email=첫번째 파라미터(email);
+public interface UserRepository extends JpaRepository<User, Long> {
+	// SELECT * FROM user WHERE username=첫번째 파라미터(username);
 	
 	Optional<User> findByEmail(String email);
